cms_incremental_pipeline/requirements.txt
# CMS Incremental Pipeline - Python Dependencies

# Web Framework
flask>=2.3.0

# HTTP Requests
requests>=2.31.0
tenacity>=8.2.0

# HTML Parsing
beautifulsoup4>=4.12.0

# Environment
python-dotenv>=1.0.0

# BigQuery
google-cloud-bigquery>=3.13.0
google-auth>=2.23.0

# Supabase (state tracking)
supabase>=2.0.0

# Dev/Testing (optional)
pytest>=7.4.0
black>=23.0.0
